refactor and fixups
