+++ mlos_bench/mlos_bench/os_environ.py
@@ -20,9 +20,18 @@ import sys
 if sys.platform == 'win32':
     import nt   # type: ignore[import-not-found]
 
+if sys.version_info >= (3, 10):
+    from typing import TypeAlias
+else:
+    from typing_extensions import TypeAlias
+
+if sys.version_info >= (3, 9):
+    ENVIRON_TYPE: TypeAlias = os._Environ[str]
+else:
+    ENVIRON_TYPE: TypeAlias = os._Environ
 
 # Handle case sensitivity differences between platforms.
 # https://stackoverflow.com/a/19023293
-environ: os._Environ[str] = nt.environ if sys.platform == 'win32' else os.environ
+environ: ENVIRON_TYPE = nt.environ if sys.platform == 'win32' else os.environ
 
 __all__ = ['environ']
